+++ docs/docs/utils/array/groupBy.md
@@ -1,6 +1,8 @@
 # groupBy 
 
-배열의 요소들을 제공된 콜백 함수 `callbackFn`에 따라 그룹화하여, 각 키에 해당하는 항목들의 배열을 포함하는 객체를 반환합니다.
+배열을 주어진 기준에 따라 그룹화합니다.
+ 
+`iteratee` 함수를 전달하여 반환된 키를 기준으로 항목들을 그룹화합니다. 각 키는 그룹화된 항목 배열을 포함하는 객체의 속성으로 할당됩니다.
 
 ## Code 
 
@@ -12,24 +14,25 @@
 
 |이름|hz|mean|성능|
 |------|---|---|---|
-|modern-kit/groupBy|20,066.27|0.0498|`fastest`|
-|lodash/groupBy|7,716.57|0.1296|`slowest`|
+|modern-kit/groupBy|6,982,132.03|0.0001|`fastest`|
+|lodash/groupBy|4,223,901.69|0.0002|`slowest`|
 
 - **modern-kit/groupBy**
-  - `2.60x` faster than lodash/groupBy
+  - `1.65x` faster than lodash/groupBy
 
 ## Interface
 ```ts title="typescript"
 function groupBy<T, K extends PropertyKey>(
-  items: T[] | readonly T[],
-  callbackFn: (item: T) => K
-): Record<K, T[]>
+  arr: T[] | readonly T[],
+  iteratee: (item: T) => K
+): Record<K, T[]>;
 ```
 
 ## Usage
 
 ```ts title="typescript"
 import { groupBy } from '@modern-kit/utils';
+
 const items = [
     { category: 'fruit', name: 'apple' },
     { category: 'fruit', name: 'banana' },
@@ -39,14 +42,14 @@ const items = [
 ];
 const group = groupBy(items, (item) => item.category);
 // {
-//     fruit: [
-//         { category: 'fruit', name: 'apple' },
-//         { category: 'fruit', name: 'banana' },
-//         { category: 'fruit', name: 'pear' },
-//     ],
-//     vegetable: [
-//         { category: 'vegetable', name: 'carrot' },
-//         { category: 'vegetable', name: 'broccoli' },
-//     ],
+//   fruit: [
+//       { category: 'fruit', name: 'apple' },
+//       { category: 'fruit', name: 'banana' },
+//       { category: 'fruit', name: 'pear' },
+//   ],
+//   vegetable: [
+//       { category: 'vegetable', name: 'carrot' },
+//       { category: 'vegetable', name: 'broccoli' },
+//   ],
 // };
 ```
