+++ packages/utils/src/array/index.ts
@@ -13,6 +13,7 @@ export * from './flatten';
 export * from './flattenDeep';
 export * from './flattenDeepThenMap';
 export * from './forEachRight';
+export * from './groupBy';
 export * from './intersection';
 export * from './mapRight';
 export * from './partition';
